vi: multi-phase refactor — fix show_changed(), extract modules, improve naming
Phase 1: Fix options show_changed() to include all 19 boolean and all
value options (was missing 13 booleans and 8 value options).

Phase 2: Replace last_substitution raw tuple with named LastSubstitution
struct, consolidate duplicate test helpers, make is_safe_exrc and
is_same_file_as_home_exrc static (no &self).

Phase 3: Rename ExResult::Message → StatusMessage, ExResult::Output →
CommandOutput for clarity on suppression semantics.

Phase 4: Extract config helpers (is_safe_exrc, is_same_file_as_home_exrc)
to new editors/vi/config.rs module.

Phase 5: Extract 15 execute_ex_* methods (~500 lines) to new
editors/vi/ex/executor.rs via split impl Editor block, reducing
editor.rs from 3574 to 3075 lines.
