+++ editors/vi/editor.rs
@@ -3,7 +3,7 @@
 //! This module contains the Editor struct which ties together all
 //! the components of the vi editor.
 
-use crate::buffer::{Buffer, BufferMode, Line, Position, Range};
+use crate::buffer::{Buffer, BufferMode, Position, Range};
 use crate::command::CommandParser;
 use crate::error::{Result, ViError};
 use crate::ex::command::SubstituteFlags;
@@ -18,6 +18,17 @@ use crate::shell::ShellExecutor;
 use crate::ui::{Screen, Terminal, TerminalSize};
 use crate::undo::UndoManager;
 use std::path::PathBuf;
+
+/// Stores the last substitution for & and :& commands.
+#[derive(Debug, Clone)]
+pub struct LastSubstitution {
+    /// The search pattern.
+    pub pattern: String,
+    /// The replacement string.
+    pub replacement: String,
+    /// Whether the substitution was global.
+    pub global: bool,
+}
 
 /// Stores the last f/F/t/T find command for ; and , repeat.
 #[derive(Debug, Clone, Copy)]
@@ -79,7 +90,7 @@ pub enum ExInsertMode {
 /// The main editor state.
 pub struct Editor {
     /// Current buffer.
-    buffer: Buffer,
+    pub(crate) buffer: Buffer,
     /// Editor mode.
     mode: Mode,
     /// Command parser for normal mode.
@@ -89,17 +100,17 @@ pub struct Editor {
     /// Ex command line input.
     ex_input: String,
     /// Editor options.
-    options: Options,
+    pub(crate) options: Options,
     /// Register storage.
-    registers: Registers,
+    pub(crate) registers: Registers,
     /// Undo manager.
     undo: UndoManager,
     /// Search state.
     search: SearchState,
     /// File manager.
-    files: FileManager,
+    pub(crate) files: FileManager,
     /// Terminal interface.
-    terminal: Terminal,
+    pub(crate) terminal: Terminal,
     /// Screen renderer.
     screen: Screen,
     /// Status/error message.
@@ -107,7 +118,7 @@ pub struct Editor {
     /// Whether message is an error.
     is_error: bool,
     /// Marks (a-z for user, other for special).
-    marks: [Option<Position>; 26],
+    pub(crate) marks: [Option<Position>; 26],
     /// Whether editor should quit.
     should_quit: bool,
     /// Exit code.
@@ -119,11 +130,11 @@ pub struct Editor {
     /// Last f/F/t/T command for ; and , repeat.
     last_find: Option<FindCommand>,
     /// Last substitution for & command.
-    last_substitution: Option<(String, String, bool)>,
+    pub(crate) last_substitution: Option<LastSubstitution>,
     /// Last command for . (dot) repeat.
     last_command: Option<LastCommand>,
     /// Last macro register for @@ repeat.
-    last_macro_register: Option<char>,
+    pub(crate) last_macro_register: Option<char>,
     /// Ex text input mode (:a, :i, :c).
     ex_insert_mode: Option<ExInsertMode>,
     /// Accumulated text for ex insert mode.
@@ -1617,7 +1628,7 @@ impl Editor {
     }
 
     /// Execute an ex command input string.
-    fn execute_ex_input(&mut self, input: &str) -> Result<()> {
+    pub(crate) fn execute_ex_input(&mut self, input: &str) -> Result<()> {
         // Handle search patterns
         if let Some(pattern) = input.strip_prefix('/') {
             self.search.update_options(&self.options);
@@ -1663,13 +1674,13 @@ impl Editor {
             ExResult::Edit(path) => {
                 self.open(&path)?;
             }
-            ExResult::Message(msg) => {
+            ExResult::StatusMessage(msg) => {
                 self.set_message(&msg);
             }
             ExResult::Error(msg) => {
                 self.set_error(&msg);
             }
-            ExResult::Output(lines) => {
+            ExResult::CommandOutput(lines) => {
                 if self.ex_standalone_mode {
                     // In ex mode, print output directly to stdout
                     // Note: Output from :p, :nu, :l is NOT suppressed by -s
@@ -1740,7 +1751,9 @@ impl Editor {
                 if let Some(msg) = self.options.set(&args)? {
                     // Route query/display output through Output so it prints in ex mode
                     // (Message is for informational status, suppressed in -s mode)
-                    Ok(ExResult::Output(msg.lines().map(String::from).collect()))
+                    Ok(ExResult::CommandOutput(
+                        msg.lines().map(String::from).collect(),
+                    ))
                 } else {
                     Ok(ExResult::Continue)
                 }
@@ -1765,7 +1778,7 @@ impl Editor {
                     self.files.set_current_file(Some(PathBuf::from(name)));
                 }
                 let info = self.file_info();
-                Ok(ExResult::Message(info))
+                Ok(ExResult::StatusMessage(info))
             }
             ExCommand::Undo => {
                 self.undo.undo(&mut self.buffer)?;
@@ -1799,13 +1812,13 @@ impl Editor {
                 self.open(&path.to_string_lossy())?;
                 Ok(ExResult::Continue)
             }
-            ExCommand::Args => Ok(ExResult::Message(self.files.format_args())),
-            ExCommand::Version => Ok(ExResult::Message(format!(
+            ExCommand::Args => Ok(ExResult::StatusMessage(self.files.format_args())),
+            ExCommand::Version => Ok(ExResult::StatusMessage(format!(
                 "{} {}",
                 env!("CARGO_PKG_NAME"),
                 env!("CARGO_PKG_VERSION")
             ))),
-            ExCommand::Help => Ok(ExResult::Message(format!(
+            ExCommand::Help => Ok(ExResult::StatusMessage(format!(
                 "{}: {}",
                 env!("CARGO_PKG_NAME"),
                 env!("CARGO_PKG_DESCRIPTION")
@@ -1865,15 +1878,15 @@ impl Editor {
             }
             ExCommand::Print { range, count } => {
                 let lines = self.get_lines_for_output(&range, count)?;
-                Ok(ExResult::Output(lines))
+                Ok(ExResult::CommandOutput(lines))
             }
             ExCommand::Number { range, count } => {
                 let lines = self.get_lines_with_numbers(&range, count)?;
-                Ok(ExResult::Output(lines))
+                Ok(ExResult::CommandOutput(lines))
             }
             ExCommand::List { range, count } => {
                 let lines = self.get_lines_for_list(&range, count)?;
-                Ok(ExResult::Output(lines))
+                Ok(ExResult::CommandOutput(lines))
             }
             ExCommand::Join { range, count } => {
                 self.execute_ex_join(&range, count)?;
@@ -1912,7 +1925,7 @@ impl Editor {
                 let cwd = std::env::current_dir()
                     .map(|p| p.display().to_string())
                     .unwrap_or_else(|_| "unknown".to_string());
-                Ok(ExResult::Message(cwd))
+                Ok(ExResult::StatusMessage(cwd))
             }
             ExCommand::Mark { line, name } => {
                 self.execute_ex_mark(line, name)?;
@@ -1922,7 +1935,7 @@ impl Editor {
             ExCommand::Open { line } => Ok(ExResult::EnterOpen(line)),
             ExCommand::Z { line, ztype, count } => {
                 let output = self.execute_ex_z(line, ztype, count)?;
-                Ok(ExResult::Output(output))
+                Ok(ExResult::CommandOutput(output))
             }
             ExCommand::ShiftLeft { range, count } => {
                 self.execute_ex_shift_left(&range, count)?;
@@ -1934,7 +1947,7 @@ impl Editor {
             }
             ExCommand::LineNumber { line } => {
                 let line_num = self.execute_ex_line_number(line)?;
-                Ok(ExResult::Output(vec![line_num.to_string()]))
+                Ok(ExResult::CommandOutput(vec![line_num.to_string()]))
             }
             ExCommand::Execute { range, buffer } => {
                 self.execute_ex_execute(&range, buffer)?;
@@ -2004,7 +2017,7 @@ impl Editor {
             if let Ok(home) = std::env::var("HOME") {
                 if !home.is_empty() {
                     let exrc_path = format!("{}/.exrc", home);
-                    if self.is_safe_exrc(&exrc_path) {
+                    if crate::config::is_safe_exrc(&exrc_path) {
                         let _ = self.execute_source(&exrc_path);
                     }
                 }
@@ -2014,66 +2027,14 @@ impl Editor {
         // 2. If exrc option is set, source ./.exrc in current directory
         if self.options.exrc {
             let local_exrc = ".exrc";
-            if self.is_safe_exrc(local_exrc) && !self.is_same_file_as_home_exrc(local_exrc) {
+            if crate::config::is_safe_exrc(local_exrc)
+                && !crate::config::is_same_file_as_home_exrc(local_exrc)
+            {
                 let _ = self.execute_source(local_exrc);
             }
         }
 
         Ok(())
-    }
-
-    /// Check if a path refers to the same file as $HOME/.exrc using inode/device comparison.
-    fn is_same_file_as_home_exrc(&self, path: &str) -> bool {
-        use std::fs;
-        use std::os::unix::fs::MetadataExt;
-        use std::path::Path;
-
-        let home = match std::env::var("HOME") {
-            Ok(h) if !h.is_empty() => h,
-            _ => return false,
-        };
-
-        let home_exrc = format!("{}/.exrc", home);
-        let meta_home = match fs::metadata(&home_exrc) {
-            Ok(m) => m,
-            Err(_) => return false,
-        };
-        let meta_local = match fs::metadata(Path::new(path)) {
-            Ok(m) => m,
-            Err(_) => return false,
-        };
-
-        meta_home.dev() == meta_local.dev() && meta_home.ino() == meta_local.ino()
-    }
-
-    /// Check if an .exrc file is safe to source.
-    ///
-    /// Per POSIX, the file must be:
-    /// - Owned by the real user ID (or process has appropriate privileges)
-    /// - Not writable by group or others
-    fn is_safe_exrc(&self, path: &str) -> bool {
-        use std::fs;
-        use std::os::unix::fs::MetadataExt;
-
-        let metadata = match fs::metadata(path) {
-            Ok(m) => m,
-            Err(_) => return false, // File doesn't exist - not an error
-        };
-
-        // Check ownership: must be owned by real user ID
-        let real_uid = unsafe { libc::getuid() };
-        if metadata.uid() != real_uid {
-            return false;
-        }
-
-        // Check permissions: not writable by group or others
-        let mode = metadata.mode();
-        if (mode & 0o022) != 0 {
-            // group or other write bit set
-            return false;
-        }
-
-        true
     }
 
     /// Execute a shell command (:! or :shell).
@@ -2358,195 +2319,6 @@ impl Editor {
         Ok(lines)
     }
 
-    /// Execute :join command - join lines together.
-    fn execute_ex_join(&mut self, range: &AddressRange, count: Option<usize>) -> Result<()> {
-        let current = self.buffer.cursor().line;
-
-        // Determine the range of lines to join
-        let (start, mut end) = if range.explicit {
-            range.resolve(&self.buffer, current)?
-        } else if let Some(c) = count {
-            // No address: current line and current + count
-            (current, (current + c).min(self.buffer.line_count()))
-        } else {
-            // No address, no count: current line and next line
-            (current, (current + 1).min(self.buffer.line_count()))
-        };
-
-        // Apply count to extend the range if both range and count are specified
-        if let Some(c) = count {
-            if range.explicit {
-                end = (end + c - 1).min(self.buffer.line_count());
-            }
-        }
-
-        if start >= end || start > self.buffer.line_count() {
-            return Ok(()); // Nothing to join
-        }
-
-        // Build the joined line
-        let mut result = String::new();
-        for line_num in start..=end {
-            if let Some(line) = self.buffer.line(line_num) {
-                let content = line.content();
-                if result.is_empty() {
-                    result = content.to_string();
-                } else {
-                    // Add space and trimmed content (POSIX: discard leading spaces)
-                    let trimmed = content.trim_start();
-                    if !trimmed.is_empty() {
-                        if !result.is_empty() && !result.ends_with(' ') {
-                            result.push(' ');
-                        }
-                        result.push_str(trimmed);
-                    }
-                }
-            }
-        }
-
-        // Delete lines from end to start+1 (in reverse to preserve line numbers)
-        for line_num in (start + 1..=end).rev() {
-            self.buffer.delete_line(line_num);
-        }
-
-        // Replace the first line with the joined result
-        if let Some(line) = self.buffer.line_mut(start) {
-            *line = Line::from(result.as_str());
-        }
-
-        self.buffer.set_line(start);
-
-        let join_count = end - start;
-        self.set_message(&format!("{} lines joined", join_count + 1));
-        Ok(())
-    }
-
-    /// Execute :put command - put text from register after line.
-    fn execute_ex_put(&mut self, line: Option<usize>, register: Option<char>) -> Result<()> {
-        let target_line = line.unwrap_or_else(|| self.buffer.cursor().line);
-        let reg = register.unwrap_or('"');
-
-        let content = self.registers.get(reg).ok_or(ViError::BufferEmpty(reg))?;
-
-        // Insert lines after target
-        let lines: Vec<&str> = content.text.lines().collect();
-        for (i, line_text) in lines.iter().enumerate() {
-            self.buffer
-                .insert_line_after(target_line + i, Line::from(*line_text));
-        }
-
-        self.buffer.set_line(target_line + lines.len());
-        Ok(())
-    }
-
-    /// Execute :copy command - copy lines to destination.
-    fn execute_ex_copy(&mut self, range: &AddressRange, dest: usize) -> Result<()> {
-        let current = self.buffer.cursor().line;
-        let (start, end) = range.resolve(&self.buffer, current)?;
-
-        // Collect the lines to copy
-        let mut lines_to_copy = Vec::new();
-        for line_num in start..=end {
-            if let Some(line) = self.buffer.line(line_num) {
-                lines_to_copy.push(line.content().to_string());
-            }
-        }
-
-        // Insert after destination line
-        let insert_after = if dest == 0 { 0 } else { dest };
-        for (i, line_text) in lines_to_copy.iter().enumerate() {
-            self.buffer
-                .insert_line_after(insert_after + i, Line::from(line_text.as_str()));
-        }
-
-        self.buffer.set_line(insert_after + lines_to_copy.len());
-
-        let copy_count = end - start + 1;
-        self.set_message(&format!("{} lines copied", copy_count));
-        Ok(())
-    }
-
-    /// Execute :move command - move lines to destination.
-    fn execute_ex_move(&mut self, range: &AddressRange, dest: usize) -> Result<()> {
-        let current = self.buffer.cursor().line;
-        let (start, end) = range.resolve(&self.buffer, current)?;
-
-        // Can't move lines into themselves
-        if dest >= start && dest <= end {
-            return Err(ViError::InvalidRange(
-                "Cannot move lines into themselves".to_string(),
-            ));
-        }
-
-        // Collect the lines to move
-        let mut lines_to_move = Vec::new();
-        for line_num in start..=end {
-            if let Some(line) = self.buffer.line(line_num) {
-                lines_to_move.push(line.content().to_string());
-            }
-        }
-
-        // Delete original lines (in reverse order)
-        for line_num in (start..=end).rev() {
-            self.buffer.delete_line(line_num);
-        }
-
-        // Adjust destination if it was after the deleted lines
-        let adjusted_dest = if dest > end {
-            dest - (end - start + 1)
-        } else {
-            dest
-        };
-
-        // Insert at new location
-        let insert_after = if adjusted_dest == 0 { 0 } else { adjusted_dest };
-        for (i, line_text) in lines_to_move.iter().enumerate() {
-            self.buffer
-                .insert_line_after(insert_after + i, Line::from(line_text.as_str()));
-        }
-
-        self.buffer.set_line(insert_after + lines_to_move.len());
-
-        let move_count = end - start + 1;
-        self.set_message(&format!("{} lines moved", move_count));
-        Ok(())
-    }
-
-    /// Execute :read command - read file into buffer.
-    fn execute_ex_read(&mut self, range: &AddressRange, file: Option<&str>) -> Result<()> {
-        let current = self.buffer.cursor().line;
-        let insert_after = if range.explicit {
-            let (_, end) = range.resolve(&self.buffer, current)?;
-            end
-        } else {
-            current
-        };
-
-        let path = file
-            .map(|s| s.to_string())
-            .or_else(|| self.files.current_file().map(|p| p.display().to_string()))
-            .ok_or(ViError::NoFileName)?;
-
-        let content = std::fs::read_to_string(&path).map_err(ViError::Io)?;
-
-        let lines: Vec<&str> = content.lines().collect();
-        for (i, line_text) in lines.iter().enumerate() {
-            self.buffer
-                .insert_line_after(insert_after + i, Line::from(*line_text));
-        }
-
-        self.buffer.set_line(insert_after + lines.len());
-
-        let bytes = content.len();
-        self.set_message(&format!(
-            "\"{}\" {} lines, {} bytes",
-            path,
-            lines.len(),
-            bytes
-        ));
-        Ok(())
-    }
-
     /// Execute :global command - execute command on matching lines.
     fn execute_ex_global(
         &mut self,
@@ -2615,200 +2387,6 @@ impl Editor {
         Ok(())
     }
 
-    /// Execute :mark command - set a mark.
-    fn execute_ex_mark(&mut self, line: Option<usize>, name: char) -> Result<()> {
-        let target_line = line.unwrap_or_else(|| self.buffer.cursor().line);
-
-        if !name.is_ascii_lowercase() {
-            return Err(ViError::MarkNotSet(name));
-        }
-
-        let idx = (name as u8 - b'a') as usize;
-        self.marks[idx] = Some(Position::new(target_line, 0));
-        Ok(())
-    }
-
-    /// Execute :z command - adjust window display.
-    fn execute_ex_z(
-        &mut self,
-        line: Option<usize>,
-        ztype: Option<char>,
-        count: Option<usize>,
-    ) -> Result<Vec<String>> {
-        let scroll = self.options.scroll;
-        let count = count.unwrap_or(2 * scroll);
-        let mut target_line = line.unwrap_or_else(|| self.buffer.cursor().line);
-
-        // If no type and no line, advance to next line
-        if line.is_none() && ztype.is_none() {
-            target_line = (target_line + 1).min(self.buffer.line_count());
-        }
-
-        // Adjust target based on type
-        let start_line = match ztype {
-            Some('+') => target_line,
-            Some('-') => target_line.saturating_sub(count - 1).max(1),
-            Some('.') | Some('=') => {
-                // Center on this line
-                let half = count / 2;
-                target_line.saturating_sub(half).max(1)
-            }
-            Some('^') => target_line.saturating_sub(2 * count - 1).max(1),
-            None => target_line,
-            _ => target_line,
-        };
-
-        // Collect lines to output
-        let mut output = Vec::new();
-        let end_line = (start_line + count - 1).min(self.buffer.line_count());
-
-        if let Some('=') = ztype {
-            // For '=', print separator line around current line
-            let cols = self.terminal.size().cols as usize;
-            let separator: String = "-".repeat(40.min(cols / 2));
-            let half = count / 2;
-            let before_start = target_line.saturating_sub(half).max(1);
-
-            for i in before_start..target_line {
-                if let Some(line) = self.buffer.line(i) {
-                    output.push(line.content().to_string());
-                }
-            }
-            output.push(separator.clone());
-            if let Some(line) = self.buffer.line(target_line) {
-                output.push(line.content().to_string());
-            }
-            output.push(separator);
-            for i in (target_line + 1)..=(target_line + half).min(self.buffer.line_count()) {
-                if let Some(line) = self.buffer.line(i) {
-                    output.push(line.content().to_string());
-                }
-            }
-        } else {
-            for i in start_line..=end_line {
-                if let Some(line) = self.buffer.line(i) {
-                    output.push(line.content().to_string());
-                }
-            }
-        }
-
-        // Update current line
-        let new_current = if ztype == Some('=') {
-            target_line
-        } else {
-            end_line
-        };
-        self.buffer.set_line(new_current);
-        self.buffer.move_to_first_non_blank();
-
-        Ok(output)
-    }
-
-    /// Execute :< command - shift lines left.
-    fn execute_ex_shift_left(&mut self, range: &AddressRange, count: Option<usize>) -> Result<()> {
-        let (start, end) = self.resolve_range(range)?;
-        let shift_amount = count.unwrap_or(1) * self.options.shiftwidth;
-
-        for line_num in start..=end {
-            if let Some(line) = self.buffer.line(line_num) {
-                let content = line.content().to_string();
-                // Count leading whitespace
-                let leading: usize = content
-                    .chars()
-                    .take_while(|c| c.is_whitespace())
-                    .map(|c| if c == '\t' { self.options.tabstop } else { 1 })
-                    .sum();
-
-                if leading > 0 {
-                    let new_indent = leading.saturating_sub(shift_amount);
-                    let trimmed = content.trim_start();
-                    let new_content = format!("{}{}", " ".repeat(new_indent), trimmed);
-                    let _ = self.buffer.replace_line(line_num, &new_content);
-                }
-            }
-        }
-
-        self.buffer.set_line(end);
-        self.buffer.move_to_first_non_blank();
-        Ok(())
-    }
-
-    /// Execute :> command - shift lines right.
-    fn execute_ex_shift_right(&mut self, range: &AddressRange, count: Option<usize>) -> Result<()> {
-        let (start, end) = self.resolve_range(range)?;
-        let shift_amount = count.unwrap_or(1) * self.options.shiftwidth;
-
-        for line_num in start..=end {
-            if let Some(line) = self.buffer.line(line_num) {
-                let content = line.content().to_string();
-                // Don't shift empty lines
-                if !content.is_empty() {
-                    let new_content = format!("{}{}", " ".repeat(shift_amount), content);
-                    let _ = self.buffer.replace_line(line_num, &new_content);
-                }
-            }
-        }
-
-        self.buffer.set_line(end);
-        self.buffer.move_to_first_non_blank();
-        Ok(())
-    }
-
-    /// Execute := command - print line number.
-    fn execute_ex_line_number(&self, line: Option<usize>) -> Result<usize> {
-        // Default to last line in buffer
-        let line_num = line.unwrap_or_else(|| self.buffer.line_count());
-        Ok(line_num)
-    }
-
-    /// Execute :@ or :* command - execute buffer contents as ex commands.
-    fn execute_ex_execute(&mut self, range: &AddressRange, buffer: Option<char>) -> Result<()> {
-        // Get the buffer to execute
-        let buffer_char = buffer.unwrap_or_else(|| {
-            // Use last executed buffer, default to unnamed
-            self.last_macro_register.unwrap_or('"')
-        });
-
-        // Get buffer contents
-        let content = if buffer_char == '"' {
-            // Unnamed buffer
-            self.registers.get('"').map(|r| r.text.clone())
-        } else if buffer_char.is_ascii_alphabetic() {
-            self.registers.get(buffer_char).map(|r| r.text.clone())
-        } else {
-            None
-        };
-
-        let content = match content {
-            Some(c) if !c.is_empty() => c,
-            _ => return Err(ViError::BufferEmpty(buffer_char)),
-        };
-
-        // Remember this buffer for @@ / **
-        self.last_macro_register = Some(buffer_char);
-
-        // Execute for each line in the range (or just once if no explicit range)
-        let (start, end) = if range.explicit {
-            self.resolve_range(range)?
-        } else {
-            let current = self.buffer.cursor().line;
-            (current, current)
-        };
-
-        for line_num in start..=end {
-            self.buffer.set_line(line_num);
-            // Execute each line of the buffer content as an ex command
-            for cmd_line in content.lines() {
-                let cmd_line = cmd_line.trim();
-                if !cmd_line.is_empty() {
-                    self.execute_ex_input(cmd_line)?;
-                }
-            }
-        }
-
-        Ok(())
-    }
-
     /// Execute :suspend or :stop command - suspend the editor.
     fn execute_ex_suspend(&mut self) -> Result<()> {
         // Only restore terminal if we're in visual mode (raw mode active)
@@ -2829,22 +2407,6 @@ impl Editor {
         }
 
         Ok(())
-    }
-
-    /// Execute :& command - repeat last substitute.
-    fn execute_ex_repeat_substitute(
-        &mut self,
-        range: &AddressRange,
-        flags: &SubstituteFlags,
-    ) -> Result<()> {
-        // Get last substitute pattern and replacement
-        let (pattern, replacement) = match &self.last_substitution {
-            Some((p, r, _)) => (p.clone(), r.clone()),
-            None => return Err(ViError::NoPreviousSubstitution),
-        };
-
-        // Use provided flags or default to last flags
-        self.substitute(range, &pattern, &replacement, flags)
     }
 
     /// Search for next match.
@@ -2884,7 +2446,7 @@ impl Editor {
     }
 
     /// Perform substitution.
-    fn substitute(
+    pub(crate) fn substitute(
         &mut self,
         range: &AddressRange,
         pattern: &str,
@@ -2930,73 +2492,12 @@ impl Editor {
         } else {
             self.set_message(&format!("{} substitutions", total_count));
             // Save for & command
-            self.last_substitution =
-                Some((pattern.to_string(), replacement.to_string(), flags.global));
+            self.last_substitution = Some(LastSubstitution {
+                pattern: pattern.to_string(),
+                replacement: replacement.to_string(),
+                global: flags.global,
+            });
         }
-
-        Ok(())
-    }
-
-    /// Resolve address range to line numbers.
-    fn resolve_range(&self, range: &AddressRange) -> Result<(usize, usize)> {
-        range.resolve(&self.buffer, self.buffer.cursor().line)
-    }
-
-    /// Execute ex delete command (:d).
-    fn execute_ex_delete(
-        &mut self,
-        range: &AddressRange,
-        register: Option<char>,
-        count: Option<usize>,
-    ) -> Result<()> {
-        use crate::command::delete;
-
-        let (start, end) = self.resolve_range(range)?;
-        let end = if let Some(c) = count {
-            (start + c - 1).min(self.buffer.line_count())
-        } else {
-            end
-        };
-
-        let start_pos = Position::new(start, 0);
-        let end_pos = Position::new(end, 0);
-        let del_range = Range::lines(start_pos, end_pos);
-
-        let result = delete(&mut self.buffer, del_range, &mut self.registers, register)?;
-        self.buffer.set_cursor(result.cursor);
-
-        let line_count = end - start + 1;
-        if line_count > 1 {
-            self.set_message(&format!("{} lines deleted", line_count));
-        }
-
-        Ok(())
-    }
-
-    /// Execute ex yank command (:y).
-    fn execute_ex_yank(
-        &mut self,
-        range: &AddressRange,
-        register: Option<char>,
-        count: Option<usize>,
-    ) -> Result<()> {
-        use crate::command::yank;
-
-        let (start, end) = self.resolve_range(range)?;
-        let end = if let Some(c) = count {
-            (start + c - 1).min(self.buffer.line_count())
-        } else {
-            end
-        };
-
-        let start_pos = Position::new(start, 0);
-        let end_pos = Position::new(end, 0);
-        let yank_range = Range::lines(start_pos, end_pos);
-
-        let _ = yank(&self.buffer, yank_range, &mut self.registers, register);
-
-        let line_count = end - start + 1;
-        self.set_message(&format!("{} lines yanked", line_count));
 
         Ok(())
     }
@@ -3304,13 +2805,13 @@ impl Editor {
     }
 
     /// Set a message.
-    fn set_message(&mut self, msg: &str) {
+    pub(crate) fn set_message(&mut self, msg: &str) {
         self.message = Some(msg.to_string());
         self.is_error = false;
     }
 
     /// Set an error message.
-    fn set_error(&mut self, msg: &str) {
+    pub(crate) fn set_error(&mut self, msg: &str) {
         self.message = Some(msg.to_string());
         self.is_error = true;
     }
@@ -3467,16 +2968,16 @@ impl Editor {
 
     /// Repeat last substitution (& command).
     fn repeat_substitution(&mut self) -> Result<()> {
-        if let Some((pattern, replacement, global)) = self.last_substitution.clone() {
+        if let Some(last_sub) = self.last_substitution.clone() {
             let flags = SubstituteFlags {
-                global,
+                global: last_sub.global,
                 confirm: false,
                 print: false,
                 count: false,
                 ignore_case: self.options.ignorecase,
             };
             let range = AddressRange::current();
-            self.substitute(&range, &pattern, &replacement, &flags)?;
+            self.substitute(&range, &last_sub.pattern, &last_sub.replacement, &flags)?;
         } else {
             self.set_error("No previous substitution");
         }
